reformat code, reload function
